+++ VoxelGame/Logic/Blocks/Core/CustomModelBlock.cs
@@ -48,7 +48,7 @@ namespace VoxelGame.Logic.Blocks
             vertCount = (uint)(blockModel.VertexCount);
         }
 
-        public override uint GetMesh(BlockSide side, byte data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
+        public override uint GetMesh(BlockSide side, uint data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
         {
             vertices = this.vertices;
             textureIndices = texIndices;
@@ -74,7 +74,7 @@ namespace VoxelGame.Logic.Blocks
             }
         }
 
-        internal override void BlockUpdate(int x, int y, int z, byte data, BlockSide side)
+        internal override void BlockUpdate(int x, int y, int z, uint data, BlockSide side)
         {
             if (side == BlockSide.Bottom && !(Game.World.GetBlock(x, y - 1, z, out _) ?? Block.Air).IsSolidAndFull)
             {
