+++ VoxelGame/Logic/Blocks/Nature/DirtBlock.cs
@@ -25,7 +25,7 @@ namespace VoxelGame.Logic.Blocks
         {
         }
 
-        internal override void RandomUpdate(int x, int y, int z, byte data)
+        internal override void RandomUpdate(int x, int y, int z, uint data)
         {
             // Check if this block can be grass
             if ((Game.World.GetBlock(x, y + 1, z, out _) ?? Block.Air).IsSolidAndFull)
